dataset/requirements_var.txt
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
clip==0.2.0
# contourpy==1.2.1
cycler==0.12.1
datasets==2.19.1
diffusers==0.27.2
dill==0.3.8
fairscale==0.4.13
ffmpeg-python==0.2.0
filelock==3.14.0
fonttools==4.52.4
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
hf_transfer==0.1.6
huggingface-hub==0.23.2
idna==3.7
ignite==1.1.0
importlib_metadata==7.1.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lmdb==1.4.1
MarkupSafe==2.1.5
# matplotlib==3.9.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
# networkx#==3.3
# numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
openai-clip==1.0.1
opencv-python==4.10.0.84
packaging==24.0
# pandas==2.2.2
pillow==10.2.0
psutil==5.9.8
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.2
safetensors==0.4.3
# scikit-learn==1.5.0
# scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.0
six==1.16.0
soupsieve==2.5
sympy==1.12
threadpoolctl==3.5.0
timm==1.0.3
tokenizers==0.19.0
torch==2.3.0
torchvision==0.16.1
tqdm==4.66.4
transformers==4.41.1
triton==2.3.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
wrapt==1.16.0
xformers==0.0.26.post1
xxhash==3.4.1
yarl==1.9.4
zipp==3.18.2